offers `Quality` field parse error (must be `double`, not `uint`)

+++ Xrpl.C/Xrpl/Client/Models/Methods/AccountOffers.cs
@@ -44,7 +44,7 @@ namespace Xrpl.Client.Models.Methods
         public Currency TakerPays { get; set; }
 
         [JsonProperty("quality")]
-        public uint Quality { get; set; }
+        public double Quality { get; set; }
 
         [JsonProperty("expiration")]
         [JsonConverter(typeof(RippleDateTimeConverter))]
